requirements.txt
openai
lancedb
docling
pyarrow
tiktoken
python-dotenv
tenacity
ipywidgets
ipython
